backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9
crewai>=0.22.5
langchain-google-genai>=0.0.9
google-generativeai>=0.4.0
python-dotenv>=1.0.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
arxiv>=2.1.0
pypdf>=4.1.0
python-docx>=1.1.0
python-pptx>=0.6.23
pyyaml>=6.0.1
marked>=1.0.0
html2pdf>=0.0.1
